pyproject.toml
[project]
name = "novelty-wealth-news-bot"
version = "0.1.0"
description = "Fetch, filter, summarize, and send Indian finance news/content ideas to Slack."
requires-python = ">=3.11"
dependencies = [
  "feedparser>=6.0.11",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "openai>=1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
